refactor: change project name to makelist
