+++ README.md
@@ -1,4 +1,4 @@
-# Payroll
+# MakeList
 
 This repository has been created as a learning tool for getting to grips with spring boot.
