+++ pom.xml
@@ -9,9 +9,9 @@
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
 	<groupId>ch.etml-es</groupId>
-	<artifactId>payroll</artifactId>
+	<artifactId>MakeList</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
-	<name>payroll</name>
+	<name>MakeList</name>
 	<description>Training depot for Spring</description>
 	<properties>
 		<java.version>17</java.version>
