+++ src/main/java/ch/etmles/makelist/Controllers/EmployeeNotFoundAdvice.java
@@ -1,4 +1,4 @@
-package ch.etmles.payroll.Controllers;
+package ch.etmles.makelist.Controllers;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ControllerAdvice;
