+++ src/main/resources/application.properties
@@ -1,1 +1,1 @@
-spring.application.name=payroll
+spring.application.name=makelist
