+++ src/test/java/ch/etmles/makelist/MakelistApplicationTests.java
@@ -1,10 +1,10 @@
-package ch.etmles.payroll;
+package ch.etmles.makelist;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class PayrollApplicationTests {
+class MakelistApplicationTests {
 
 	@Test
 	void contextLoads() {
